Add standalone.xml configuration
Added standalone.xml configuration for JBoss Wildfly with specific modules and subsystems. The configuration allows to tune up server settings and functionality for convenience and performance. Additionally, jboss-deployment-structure.xml was added to exclude specific modules during the deployment, improving the deployment time and system resources. Changes in docker-compose.dev.yml and persistence.xml are commented out, preserving it as is for further reference.